fix: explicitly use int64_t instead of long long is apparently 32 bit in Windows, long long is 64. So we explicitly use int64_t everywhere as the default type.

+++ objects/core.cpp
@@ -16,7 +16,7 @@ Value next_core_clock(const Value *args, int numargs) {
 	// TODO: clock_t is 64bits long, casting it to 32bits will reset the time
 	// every 36 mins or so. Casting it to double will probably result losing
 	// some precision.
-	return Value(clock());
+	return Value((double)clock());
 }
 
 Value next_core_type_of(const Value *args, int numargs) {
